+++ projects/core/src/main/kotlin/delegates-jvm.kt
@@ -26,15 +26,16 @@ import nl.komponents.kovenant.Kovenant
 import nl.komponents.kovenant.Promise
 import nl.komponents.kovenant.async
 import java.util.concurrent.atomic.AtomicInteger
+import kotlin.reflect.KProperty
 
 
-internal class LazyPromise<T>(
+public class LazyPromise<T>(
         //Need to allow `null` context because we could easily
         //create this property before Kovenant gets configured.
         //that would lead to this property using another Context
         //than the rest of the program.
         private val context: Context?,
-        initializer: () -> T) : Lazy<Promise<T, Exception>>() {
+        initializer: () -> T) /*: Lazy<Promise<T, Exception>>() */{
 
 
     private @Volatile var initializer: (() -> T)?
@@ -45,8 +46,13 @@ internal class LazyPromise<T>(
         this.initializer = initializer
     }
 
+    @Suppress("UNUSED_PARAMETER")
+    operator fun getValue(thisRef: Any?, property: KProperty<*>): Promise<T, Exception> = initOrGetPromise()
+/*
+    //See if Lazy will be openend up again
     override val value: Promise<T, Exception> get() = initOrGetPromise()
     override fun isInitialized(): Boolean = promise != null
+*/
 
     private fun initOrGetPromise(): Promise<T, Exception> {
         // Busy/Spin lock, expecting async to return quickly