+++ projects/incubating/src/main/kotlin/sequences.kt
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.komponents.kovenant.functional
+package nl.komponents.kovenant.incubating
 
 import nl.komponents.kovenant.*
 import java.util.*
@@ -30,12 +30,12 @@ import java.util.*
  */
 public fun <V, R> Sequence<V>.mapEach(context: Context = Kovenant.context, bind: (V) -> R): Promise<List<R>, Exception> {
     val deferred = deferred<List<R>, Exception>(context)
-    context.workerContext offer {
+    context.workerContext.offer {
         //TODO ArrayList is jvm only
         val promises = ArrayList<Promise<R, Exception>>()
         forEach {
             value ->
-            promises add async(context) { bind(value) }
+            promises.add(async(context) { bind(value) })
         }
         val masterPromise = all(promises)
         masterPromise success {